Add error handling and checking to prevent TwitchLib exception.
The Twitch exception in Question is:
"
    Unhandled Exception: TwitchLib.Client.Exceptions.BadStateException:
    Must be connected to at least one channel.
       at TwitchLib.Client.TwitchClient.GetJoinedChannel(String channel)
       at TwitchLib.Client.TwitchClient.SendMessage(String channel,
       String message, Boolean dryRun)
       at QuestionBot.Twitch.Client.SendMessage(String content)
       at QuestionBot.Twitch.Client.HandleMessageReceivedAsync(Object
       sender, OnMessageReceivedArgs e)
       at System.Threading.ExecutionContext.RunInternal(ExecutionContext
       executionContext, ContextCallback callback, Object state)
    --- End of stack trace from previous location where exception was
    thrown ---
       at System.Threading.ThreadPoolWorkQueue.Dispatch()
"

Add code to the SendMessage method that checks whether or not the client
is currently joined to any channels. Reconnect and wait a bit, if it's
not. If the client waited for 120 seconds and the client isn't
connected, return and don't send a message at all. Also log all of that.

Put the SendMessage method of the TwitchLib client called by
Client.SendMessage in a try block and log, if an exception gets caught.

The first change should hopefully prevent this error, while the second
change catches the exception, if it occurs despite the prevention
attempt.